Undo renamings
